Logic separation
Separated logic from main program file
